subscribe: remove object from subscription event
Subscriptions are deep, so it doesn't make sense to surface the same
path or instance object on which you subscribed in the callback; you
already have that context. And we can't surface arbitrary specific
objects without losing type information. So, remove the `object` from
the subscription callback and rely on the `message` to tell you what
changed.

Note that the client surfaces stable references to object instances so
in frameworks like React you can still rely on rendering updates to only
the data that changed, since reference equality is preserved when the
current data is obtained from the usual accessor methods.